requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
decorator==4.4.2
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0 
httpcore==1.0.2
httpx==0.25.2  
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.3
moviepy==1.0.3
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.2
openai==0.28.0
packaging==23.2
Pillow==10.2.0
preshed==3.0.9
proglog==0.1.10
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.4